Move StepNavHelper to correct directory
This commit correctly namespaces StepNavHelper and
also corrects the directory of the equivalent spec.
