+++ app/views/govuk_publishing_components/components/_step_by_step_nav.html.erb
@@ -26,7 +26,7 @@
         <%
           step_is_active = step_index + 1 == highlight_step
           step_count += 1
-          id = StepNavHelper.new.generate_step_nav_id(step[:title])
+          id = GovukPublishingComponents::Components::StepNavHelper.new.generate_step_nav_id(step[:title])
 
           logic = false
           logic = step[:logic] if ["and", "or"].include? step[:logic]
@@ -71,7 +71,7 @@
               options[:link_index] = 0
             %>
             <% step[:contents].each do |element| %>
-              <%= StepNavHelper.new.render_step_nav_element(element, options) %>
+              <%= GovukPublishingComponents::Components::StepNavHelper.new.render_step_nav_element(element, options) %>
               <%
                 if element[:type] == 'list'
                   options[:link_index] += element[:contents].length
